Implemented GVP for GOPATH management Modified directory structure to match Improved README installation instructions Improved shell scripts
