+++ packages/package.sh
@@ -6,8 +6,10 @@
 
 #### GLOBALS
 
-bin_location=/usr/bin/
-sample_location=/etc/distributive.d/samples/
+binary_dir=/usr/bin/
+sample_dir=/etc/distributive.d/samples/
+binary=../bin/distributive
+samples=../samples
 
 #### BUILD
 
@@ -17,6 +19,6 @@ cd -
 
 #### PACKAGE
 
-# use FPM to build a package with the correct version number
-PATH=$PATH:~/.gem/ruby/2.2.0/bin/
-fpm -s dir -t rpm -n distributive -f -d go -m langston@aster.is -v $version --epoch 0 ../distributive=$bin_location ../samples=$sample_location
+gem_dir="~/.gem/ruby/2.2.0/bin/"
+PATH=$PATH:$gem_dir
+fpm -s dir -t rpm -n distributive -f -d go -m langston@aster.is -v $version --epoch 0 $binary=$binary_dir $samples=$sample_dir
