+++ run.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 source build.sh
-sudo ./distributive -d "" -f samples/misc.json --verbosity="debug"
+sudo ./bin/distributive -d "" -f samples/systemctl-fail.json --verbosity="debug"
